Moved check on TypeCherk
